Minor performance improvements
